Fixed Map Showing Button Function
The display was working glitchy and the word was not showing up: it had
to do with the while loop and how the message screen was being run in a
way that was multiple calls per looping and that is the type of work
that was involved with. Then the display now is toggling the map and the
display is showing appropriately now.